backend/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
redis==4.5.5
prometheus-client==0.16.0
gunicorn==20.1.0